+++ resources/connector/connector.py
@@ -709,7 +709,6 @@ class Connector:
         "action": "set-models",
         "mainModel": self.coder.main_model.name,
         "weakModel": self.coder.main_model.weak_model.name,
-        "maxChatHistoryTokens": self.coder.main_model.max_chat_history_tokens,
         "reasoningEffort": self.coder.main_model.get_reasoning_effort(),
         "thinkingTokens": self.coder.main_model.get_thinking_tokens(),
         "info": info,