+++ src/renderer/src/components/common/StyledTooltip.tsx
@@ -1,13 +1,13 @@
+import { ReactNode } from 'react';
 import { Tooltip } from 'react-tooltip';
 
 type Props = {
   id: string;
+  content?: ReactNode;
 };
 
-export const StyledTooltip = ({ id }: Props) => (
-  <Tooltip
-    id={id}
-    className="!bg-neutral-900 !text-neutral-300 !text-xxs !py-1 !px-2 !opacity-100 !rounded-md !max-w-[300px] z-50"
-    border="1px solid #495057"
-  />
+export const StyledTooltip = ({ id, content }: Props) => (
+  <Tooltip id={id} className="!bg-neutral-900 !text-neutral-300 !text-xxs !py-1 !px-2 !opacity-100 !rounded-md !max-w-[300px] z-50" border="1px solid #495057">
+    {content}
+  </Tooltip>
 );