+++ src/renderer/src/components/message/MessageBlock.tsx
@@ -1,7 +1,6 @@
 import { CommandOutputMessageBlock } from './CommandOutputMessageBlock';
 import { LoadingMessageBlock } from './LoadingMessageBlock';
 import { LogMessageBlock } from './LogMessageBlock';
-import { ModelsMessageBlock } from './ModelsMessageBlock';
 import { UserMessageBlock } from './UserMessageBlock';
 import { ReflectedMessageBlock } from './ReflectedMessageBlock';
 import { ResponseMessageBlock } from './ResponseMessageBlock';
@@ -10,7 +9,6 @@ import {
   isCommandOutputMessage,
   isLogMessage,
   isLoadingMessage,
-  isModelsMessage,
   isUserMessage,
   isReflectedMessage,
   isResponseMessage,
@@ -32,10 +30,6 @@ export const MessageBlock = ({ baseDir, message, allFiles }: Props) => {
 
   if (isLogMessage(message)) {
     return <LogMessageBlock message={message} />;
-  }
-
-  if (isModelsMessage(message)) {
-    return <ModelsMessageBlock message={message} />;
   }
 
   if (isReflectedMessage(message)) {