+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - arrow up/down does not change the history prompt after editing the text anymore
 - improved agent system prompt
 - added gpt-4.1 and gpt-4.1-mini models for Agent
+- added more information about selected model (tokens, cost per 1M tokens, etc.)
 
 ## [0.9.0]